squash: Check C++ reserved keywords
Since we camelCase and/or prefix/suffix most identifiers, they're pretty
safe. The exception is namespaces, which consist of the top-level
package name as well as the leaf namespace. These strings are processed
to be roughly valid identifiers, but then also need checking if the
(processed) string is a reserved keyword.

Signed-off-by: David Feltell <david.feltell@foundry.com>