squash: Fix Windows `ctest` run
The default for CMake on Windows is to use MSVC, which is multi-config,
i.e. the generator ignores the CMAKE_BUILD_TYPE variable at configure
time. Instead, the build type becomes a build-time (or test-time) flag.

We could use Ninja, as we do elsewhere, to provide a more consistent
cross-platform configuration. This also has the advantage of
automatically using multiple cores for the build.

However, as we have discovered, using Ninja on Windows means we have to
do the `vcvarsall.bat` preamble dance to configure the environment,
which is a complication for CI.

The CMake test project here is simple enough that a multi-core build
isn't really necessary.

So take the simpler option of explicitly specifying the build type
at build/test time, as well as configure time.