squash: Duplicate convenience functions from SpecificationBase
SpecificationBase doesn't exist in C++, and we plan to remove it. A base
class is not needed except for polymorphism and reducing code,
duplication neither of which are (currently) issues for us.

Switch to using `class` rather than `struct` to be more explicit about
symbol visibility, now that we have private members. Do this for trait
classes too, so that the tree-sitter query still works (and since we'll
be similarly removing the TraitBase base class in an upcoming commit).

Signed-off-by: David Feltell <david.feltell@foundry.com>