[CI] Run `ctest` tests for C++ generated code
Closes #17. Run the `pytest` tests marked with "ctest", which involves
building and executing a CMake CTest test suite.

Use an OpenAssetIO build action, to be replaced in the future with a
Conan package, once available.

Avoid cross-platform issues, for now, by only building/testing on the
ASWF Docker image. Once the Conan package is available, this can be
converted to a per platform build matrix.

Signed-off-by: David Feltell <david.feltell@foundry.com>