[Core] C++ code generation
Closes #12. Generate output familiar to users of the Python generator,
with a few notable differences.

Each trait/specification class has it's own file, unlike in Python,
since this is fairly idiomatic in C++ and, in particular, we do not want
to  unnecessarily increase compile times for C++ users.

The package itself has a top-level hoisting header that imports
everything. At the next level, there is a header to import all the
packages traits or specifications. Below that, there is a header for
each namespace, which imports all its traits/specifications. This
broadly mirrors the Python package structure (i.e. hoisting imports in
`__init__.py`s).

All traits and specifications exist under an `inline namespace`
providing an (eventual) customisation point for versioning
trait/specification classes, in case they're used as currency across a
DSO boundary. Currently this namespace is simply "v1", configured as a
global variable.

Similarly, the ABI version of OpenAssetIO, which the generated code
depends on, is explicitly specified so that accidental use of a trait
view with a newer ABI OpenAssetIO is a linker error, rather than runtime
error.

Since we camelCase and/or prefix/suffix most identifiers, they're pretty
safe. The exception is namespaces. These strings are processed
to be roughly valid identifiers. We assume a valid identifier in Python
is also a valid identifier in C++, so can use the Python built-in
function `isidentifier`. We then also need to check if the (processed)
string is a reserved keyword. Python has a built-in function for this,
but that is not available in C++. So bundle a list of reserved C++
keywords and check against that.

The traits that a specification depends on are imported using the
package-level hoisting header, i.e. they import everything from a
package even if they don't need it, for parity with Python. Future work
will refine this for C++.

Deviate from the nomenclature used in Python generation (e.g. for
variable names, log messages) to use "namespace" rather than "module".
We do still talk about a "package", even though that is not a C++
language construct, to refer to the broader trait package being
generated.

We don't expect polymorphism to be a common use case, and can easily
duplicate methods, so having a common base class for every
trait/specification class has little benefit. So simply duplicate the
convenience functions that were in `TraitBase` and `SpecificationBase`
(noting that `SpecificationBase` never had a C++ implementation, anyway)
into every generated class. Future work will remove those base classes
from generated Python code and then remove them entirely from the
OpenAssetIO core.

With the additional complexities of C++, simply translating the
`generate` function from the Python generator becomes unwieldy, so use
a `Renderer` utility class in the C++ case. Future work will look into
whether this and other chunks of code can be better shared between
generators.

Test the C++ generated code in two ways. Firstly, the `tree-sitter`
library is used to parse the AST of the C++ code, for use in inspecting
the generated docstrings and comments. Secondly, `pytest` builds and
invokes a CMake CTest project. Doing this through `pytest` is handy
since it lets us re-use fixtures for generating the code and managing
temporary directories.

A separate `pip` requiremenst file `ctest-requirements.txt` is created
to hold the various tools and linters for running the `ctest` tests.
Keeping these separate means that valuable time on the CI can be saved,
for jobs that do not need to run the C++ tests.

These C++ tests have a pytest marker of "ctest", so that they can be
executed on their own or explicitly skipped.

The tests require a `OPENASSETIO_TRAITGENTEST_CMAKE_PRESET` environment
variable to be set to a valid CMake preset, currently "test" and "lint",
with the latter enabling linters and static and dynamic analysis. If the
environment variable is not set, then the tests will be skipped, and
`pytest` output will show "1 skipped".

Explicitly disable "ctest"-marked tests in CI. Even though `ctest` tests
are disabled unless the `OPENASSETIO_TRAITGENTEST_CMAKE_PRESET` env var
is set, being explicit makes it clearer to future devs that there is a
test that will not run. The only difference in output is that `pytest`
reports "1 deselected", as opposed to "1 skipped".

The default for CMake on Windows is to use MSVC, which is multi-config,
i.e. the generator ignores the CMAKE_BUILD_TYPE variable at configure
time. Instead, the build type becomes a build-time (or test-time) flag.

We could use Ninja, as we do elsewhere, to provide a more consistent
cross-platform configuration. This also has the advantage of
automatically using multiple cores for the build.

However, as we have discovered, using Ninja on Windows means we have to
do the `vcvarsall.bat` preamble dance to configure the environment,
which is a complication for CI.

The CMake test project here is simple enough that a multi-core build
isn't really necessary. So take the simpler option of explicitly
specifying the build type at build/test time, as well as configure time.

Signed-off-by: David Feltell <david.feltell@foundry.com>