[Tests] Fix tests now missing traits do not error
Closes #48.

In OpenAssetIO/OpenAssetIO#970 `TraitsData::getTraitProperty` was
updated such that an exception is no longer thrown when attempting to
retrieve a property of trait that is not imbued in the data.

This means the behaviour of the TraitGen view classes is now as
intended, finally.

So update the tests to reflect this.

Signed-off-by: David Feltell <david.feltell@foundry.com>