[Core] Fix bool vs. int detection in Python setters
Closes #34. Python `int` trait properties could easily be set to `bool`
values without realising, since `bool` is a subclass of `int`.

This would probably not cause any problems as long as Python was the
only language in use. However, the value would be stored in the
underlying C++ `std::variant` as a `bool` value. This means a subsequent
C++ trait view class, which was expecting an `int`, would fail to
retrieve the value and throw an exception.

So be more strict during the type check when getting or setting trait
properties.

Signed-off-by: David Feltell <david.feltell@foundry.com>