[Core] Use set for Python kTraitSet
Closes #94. Reverts #55. Revert change for Specification class's
`kTraitSet` member to `frozenset`, since this is incompatible with older
versions of pybind11.

pybind11 is used internally to manage Python<->C++ bindings in
OpenAssetIO. If pybind11 is used in a host application too, the versions
must (currently) match.

So support host applications that ship with an older pybind11 by
eschewing `frozenset`, for now.

See also OpenAssetIO/OpenAssetIO#1420.

Signed-off-by: David Feltell <david.feltell@foundry.com>