[Core] Add UI usage options to schema
Closes #103.

The design for UI delegation makes use of traits to specify the kind of
UI that the host wants the delegate to provide. These will be distinct
from other usages of traits and specifically for UI. The design includes
a uiPolicy introspection method as well as the main populateUI method.
So we need two new usage types.

So add a ui and uiPolicy usage types for traits and specifications to
the YAML schema

Signed-off-by: David Feltell <david.feltell@foundry.com>