[CI] Bump upload/download-artifact to v4
We had a mixture of v3 and v4.  v3 is no longer supported by GitHub, so
we must update.

Presumably we had to pin to v3 due to the NodeJS version required for
v4, which was not available in the old VFX22 Docker image. However, we
now use a VFX24 Docker image, which should work fine.

Signed-off-by: David Feltell <david.feltell@foundry.com>